Splitting up AllocatorTestBase.kt into multiple sets of integration tests Extracted from #6976.
